requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
numpy
scikit-learn
chromadb
torch
transformers
pydantic
python-dotenv
umap-learn
matplotlib
seaborn
pandas
tqdm
scipy
httpx
pytest
